Fix the VS Code debug configuration
The default configuration screws up paths given in the arguments list.
Because of that, we now run the fyrc binary so we can utilize the fact that it hard prepends the directory of the given program to any path and only provide the relative path.
The binary is rebuilt before every launch.