WIP: expand platform config functionality
The compiler now supports flags to specify native compilers and configurations for them.
These are currently set up and will cause errors if specified incorrectly, but are not in use.
This commit also adds the first few platform configurations. The naming scheme may be subject to change.
Currently there is no functionality to list the available configs, it will be added in a future commit.