[NEEDSREVIEW] Work around some bugs in runtime
Since not all group pointers are unset after a `Free()`, `Malloc()` could be given a group pointer to an unallocated area.
This would cause accesses to invalid addresses when `last()` was called, causing undefined behavior.
This included things like `malloc()` returning the same pointer twice, completely breaking the program.

This implementation is a hideous hack to prevent this.
It uses the lowest bit of the group pointer to define groups that should not be explicitly free'd.
These pointer would simply be invalidated without other actions.

A proper fix for this would involve clearing all pointers of a pointer list when a single group from it is free'd.